+++ infrastructure/env/get-env-from-vcap.sh
@@ -13,5 +13,4 @@ echo $VCAP_SERVICES | ./env/jq -r '
         )
         | @tsv'
 
-echo "KEYCLOAK_DATABASE_URI=$(echo $VCAP_SERVICES | ./env/jq -r '.postgres[] | select(.name == "keycloak-database") .credentials .uri')"
 echo "RAILS_DATABASE_URL=$(echo $VCAP_SERVICES | ./env/jq -r '.postgres[] | select(.name|test("psd-d*")) .credentials .uri')"