+++ psd-web/db/migrate/20181105134010_add_correspondence_to_activities.rb
@@ -2,7 +2,9 @@ class AddCorrespondenceToActivities < ActiveRecord::Migration[5.2]
   disable_ddl_transaction!
 
   def change
-    add_reference :activities, :correspondence, index: false, foreign_key: true
-    add_index :activities, :correspondence_id, algorithm: :concurrently
+    safety_assured do
+      add_reference :activities, :correspondence, index: false, foreign_key: true
+      add_index :activities, :correspondence_id, algorithm: :concurrently
+    end
   end
 end