+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2020-04-16
+- Extended the question to identify cases related to the coronavirus outbreak to:
+  - Allegations
+  - Enquiries
+  - Projects
+
+## 2020-04-15
+- Added Your account page and ability to change your name
+
+## 2020-04-07
+- Added ability to edit the coronavirus related flag
+
 ## 2020-04-03
 - Added a question for MSA users to identify cases related to the coronavirus outbreak.